@@ -11,10 +11,11 @@ export class DialogService {
   constructor(private dialog: MatDialog) { }
 
   openConfirmDialog(msg) {
-    this.dialog.open(MatConfirmDialogComponent, {
+    return this.dialog.open(MatConfirmDialogComponent, {
       width: '390px',
       panelClass: 'confirm-dialog-container',
       disableClose: true,
+      position: {top: '10px'},
       data: {
         message: msg
       }
